+++ app/src/main/res/layout/activity_movie_booking.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MovieBookingActivity">
+    tools:context=".ui.MovieBookingActivity">
 
     <ScrollView
         android:layout_width="match_parent"
