+++ app/src/main/res/layout/activity_movie_booking_check.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MovieBookingCheckActivity">
+    tools:context=".ui.MovieBookingCheckActivity">
 
     <TextView
