+++ app/src/main/res/layout/activity_movie_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".movielistactivity.MovieListActivity">
+    tools:context=".ui.movielistactivity.MovieListActivity">
 
     <ListView
         android:id="@+id/lv_movie"
